wired up a  react handlecase with redux and fixed minor bugs
