--- src/client/app/actions/admin/meterDropDown.js
@@ -1,14 +1,0 @@
-
-export const UPDATE_IMPORT_METER_ID = 'UPDATE_IMPORT_METER_ID';
-
-export function updateSelectedMeter(meterID) {
-	return { type: UPDATE_IMPORT_METER_ID, meterID };
-}
-
-export function changeSelectedMeter(meterID) {
-	return dispatch => {
-		dispatch(updateSelectedMeter(meterID));
-		return Promise().resolve();
-	};
-}
-
