+++ src/client/app/components/MeterDropDownComponent.jsx
@@ -1,17 +1,17 @@
-/**
- * Created by Eduardo on 4/12/2017.
- */
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
 import React from 'react';
 
 export default class MeterDropDownComponent extends React.Component {
 	constructor(props) {
 		super(props);
-		console.log(props);
 		this.handleMeterSelect = this.handleMeterSelect.bind(this);
 	}
 
-	handleMeterSelect(event) {
-		this.setState({ selectedMeter: event.target.value });
+	handleMeterSelect(e) {
+		this.props.updateSelectedMeter(e.target.value);
 	}
 
 	render() {
